01-04-25_SF_Agents_meetup_[Talk]/demo/demo-agent/pyproject.toml
[project]
name = "openai-agent"
version = "0.0.1"
description = "this is a example how incorporate other agent frameworks"
readme = "README.md"
authors = [{ name = "BeeAI Demo" }]
requires-python = ">=3.11,<4.0"
dependencies = [
    "acp-sdk", # version determined by beeai-sdk
    "beeai-sdk==0.0.11",
    "pydantic-settings>=2.7.1",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
    "openai-agents",
]

[project.scripts]
openai-agent = "openai_agent.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"